// src/shared/types.ts
export enum SelectedSection {
  Home = "Home",
  HomeID = "home",
  AboutUs = "About Us",
  AboutUsID = "aboutus",
  Features = "Features",
  FeaturesID = "features",
  Recipes = "Recipes",
  RecipesID = "recipes",
  TakeAway = "Take Away",
  TakeAwayID = "takeaway",
  Testimonial = "Testimonial",
  TestimonialID = "testimonial"
}